Fix PropertyLayer dtype warnings in tests

+++ tests/test_solara_viz.py
@@ -4,6 +4,8 @@ import re
 import random
 import unittest
 
+import numpy as np
+
 import ipyvuetify as vw
 import pytest
 import solara
@@ -118,7 +120,7 @@ def test_call_space_drawer(mocker):
         def __init__(self, seed=None):
             super().__init__(seed=seed)
             layer1 = PropertyLayer(
-                name="sugar", width=10, height=10, default_value=10.0
+                name="sugar", width=10, height=10, default_value=np.float64(10.0)
             )
             self.grid = MultiGrid(
                 width=10, height=10, torus=True, property_layers=layer1
